requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==11.0.0
numpy==1.26.4
nudenet==3.4.2
httpx==0.27.2
pydantic==2.9.2
tensorflow==2.18.0
keras==3.5.0
torch==2.3.0
transformers==4.39.3
tf-keras
